Update card summary design (#255)
Updates the summary row of the card header.

- Removes "Include" for NA/Inf
- Move NA/Inf contents further to the right of the card summary.
- NA/Inf displays "NA/Inf" with a check or X. Check if yes, X if no. If
no NA or Inf are present in the column, nothing is displayed
- For value/selected portion of summary, the text is automatically
truncated depending on screen width.
- For categorical variables, the selected/value display is as follows
- if the total length of "level 1, level 2, level 3" is more than 40
characters, the text "N levels selected" is used.
  - otherwise, the selected levels are displayed
- 40 characters is a magic number and is an approximation. I found it to
be ~300px which is the total text limit. But different font sizes,
capitalization vs. not, browser zoom, etc. will all affect the actual
number of characters that = Xpx. So there's never going to be a "right"
value for this
- For numeric and dates, an endash is used instead of "-" to separate
the selected values
- For numeric variables, if there are > 4 significant digits the number
is formatted in scientific notation with 4 significant digits. This is
related to #251. If the current implementation is not exactly what we
want, it should affect this PR. We can refine the method with #251.
- The font size of some peripherals of input elements were adjusted to
match the font size in the card header.

Fixes #220