280 locked filter (#297)
Closes #280 

Adds a `locked` property to `FilterState` class. A locked filter state
cannot be disabled or removed but can still be modified. `locked` and
`fixed` properties can be set independently of one another.

Changes to `FilterState` class:
* `locked` argument is added to constructors and wrappers; the argument
receives a logical flag and defaults to `FALSE`
* a private field `locked` is added to `FilterState` to store the
property
* `$get_state` is modified to return the `locked` value in addition to
all other filter state properties
* `$set state` is modified to interrupt an attempt at disabling a filter
state that is locked
* `$ui` is modified to only display the disable and remove buttons if
`private$locked` is `FALSE`
* `$ui` is modified to display an additional icon if `private$locked` is
`TRUE`
* Analogous changes are made to `FilterStateExpr` class.
* The `locked` argument is also added to `filter_var`.
* Modified `FilterStates$remove_filter_state` and
`FilterStates$clear_filter_states` so that locked filter states are
omitted.
* The `lock` icon has been transferred from the `fixed` property to the
`locked` property. A fixed state thus needs a new icon. `burst` is used
as a placeholder but a suitable replacement must be found.
* Unit tests were amended.