428 get call bug (#430)
Fixes #428 

Having `$get_call` return `FALSE` in some circumstances caused errors
downstream. Specifically, `calls_combine_by` demands that all arguments
be calls, logical vectors are not allowed.

`$get_call` would return `FALSE` when no values are selected, including
NAs, meaning that no observations are to be included. This change was
introduced to [prettify filtering calls in PR
402](https://github.com/insightsengineering/teal.slice/pull/402).

In this PR the call for the latter case has been modified to be
explicit. For available choices of `c("a", "b", "c")`:
`!var %in% c("a", "b", "c")` with no NAs in data
`is.na(var) | !var %in% c("a", "b", "c")` with NAs present, included
`!is.na(var) & !var %in% c("a", "b", "c")` with NAs present, excluded

Admittedly, this is cumbersome and somewhat goes against the original
work in #402. However, I believe this is the way to go because
no-selection filter calls will be rare and this unseemly sight will sore
users' eyes less than the old all-inclusive filter calls.