Prevent to deselect any value in DateFilterState 429 (#433)
this fixes: https://github.com/insightsengineering/teal.slice/issues/429

I have updated the DateFilterState and incorporated restoration features
for previous selections. Each start and end date is now individually
assessed for more accurate notifications.

The use of shinyWidgets::airDatepickerInput in DateTimefilterState poses
a challenge, as it lacks a return value when cleared. This requires
modifications to enable the observation of user actions and the
reversion to the previous selection.

Demo


https://github.com/insightsengineering/teal.slice/assets/6700955/17b4e04a-2e8c-4518-a05e-522255d251ec

---------

Co-authored-by: kartikeya <kartikeya.kirar@unicle.life>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>