add disabling and reenabling of rewind and reset buttons (#444)
Fixes #443 

When state history changes, the rewind and reset buttons are hidden if
history only contains the current state. The hiding is delayed because
otherwise the buttons disappear for a moment when history is updated.
This creates the possibility of the rewind button being clicked again
during that delay window.

In this PR the buttons are disabled for the duration of the delay.