The RangeFilterState's x-axis tick labels are somewhat obstructed. (#442)
this fixes https://github.com/insightsengineering/teal.slice/issues/436


This problem doesn't occur in every system. While I managed to replicate
it on a Mac, I couldn't reproduce it on my Windows computer. I'm
somewhat uncertain whether this discrepancy is due to screen size or the
operating system, but it appears that adjusting the` ticklen` helps
resolve it. If the problem persists on other systems, we could attempt
to address it by setting `ticks = "inside"` within Plotly's
`plot_layout.`

Co-authored-by: kartikeya <kartikeya.kirar@unicle.life>