415 range selector colors (#456)
Fixes #451 

The error resulted from adding the second histogram to the plot, which
is based on the filtered data. All data was filtered out, `x_reactive`
returnes `integer(0)`, which results in all bins being colored.

Here, an `if` statement is added when creating the second histogram
trace, such that the trace itself is `NULL` if there are no data.