handle unused levels and "" choice (#587)
Fixes #586 

1. As noted
[here](https://github.com/insightsengineering/teal.slice/issues/586#issuecomment-2093746177)
having "" in the vector of names is problematic as `x[<choices>]` would
return `NA` for this item. This is why I switched to `x[match(<choices>,
names(x))]]`
2. `droplevels(x)` in the beginning of the constructor to not bother
ChoicesFilterState with unused factor levels during a time when the
class is working. It used to be this way so that zero-count-choices are
never shown in the filter-state-card.

---------

Co-authored-by: 27856297+dependabot-preview[bot]@users.noreply.github.com <27856297+dependabot-preview[bot]@users.noreply.github.com>
Co-authored-by: vedhav <vedhaviyash4@gmail.com>